requirements.txt
praw~=7.7.0
openai~=0.27.2
Flask~=2.0.2
requests~=2.26.0
python-dotenv~=0.19.2
invoke~=2.0.0
SQLAlchemy~=2.0.8
alembic~=1.10.2